Fixes found on Ubuntu linux

+++ Tests/poller.cpp
@@ -136,6 +136,6 @@ static TestSuite ts("io::poller", {
         rewind(fg.file());
 
         th.join();
-        KSS_ASSERT(d.numRead() == mydata.size() * 5);
+        KSS_ASSERT(d.numRead() == static_cast<ssize_t>(mydata.size() * 5));
     })
 });
